certificate_cheker_2_1/requirements.txt
certifi~=2023.7.22
python-dateutil~=2.9.0.post0
pytest~=7.4.4
requests~=2.31.0
urllib3
python-dotenv~=1.0.1